Move author base to init to avoid circular imports
